extract modal to separte component
